Refactors CarouselImageViewer Widget places it in Core directory, CarouselImageViewer can accept Url list or File list. and Bool to enable Close functionality
